requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
pyee==11.1.1
slackclient==2.9.4
slackeventsapi==3.0.3
typing_extensions==4.12.2
Werkzeug==3.0.4
yarl==1.11.1
